test: fix imports for the moved test files
